poi/src/requirements.txt
alembic==1.12.0
async-timeout==4.0.3
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
mysql-connector-python==8.2.0
flask_cors==4.0.0
et-xmlfile==1.1.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
cffi==1.16.0
confluent-kafka==2.3.0
cryptography==43.0.0
pika==1.3.2
pycparser==2.22
Faker==21.0.0
psutil==6.0.0